+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## Unreleased
+## 4.3.0 - 2026-03-18
 
 - **BC Break!**: Moved the compression property to the `PdfWriter` class.
 - **BC Break!**: Moved all output functions to the new `PdfWriter` class.